refactor: ProductList 생성 기능을 ProductService로 이동
ProductList를 생성하는 코드를 Product 도메인에서 ProductService단으로 이동시켰습니다.
